requirements.txt
pillow
click
flask
gevent
